+++ plugins/gh-datainmap/CHANGELOG.md
@@ -5,6 +5,7 @@
 - De zichtbare locaties worden nu ook in een lijstweergave getoond. Dit is een nieuw onderdeel welke apart gestyled dient te worden (voorzien in meegeleverde CSS).
 - Zoekfunctie weer zichtbaar in beheeromgeving.
 - De kaart omgezet naar een landmark zodat deze eenvoudiger gevonden kan worden.
+- Ref gebruik in kaart omgezet van named refs naar React.createRef().
 
 **1.6.5 (2020-07-21)**